Store base configuration in demo.
In some previous change, the drm configuration method was changed to
reset the player's configuration first, to remove any previous asset
drm configuration.
However, this also removed any user-originating configuration.
This changes the demo to store a record of the configuration from the
perspective of the user, and re-applies that instead to "reset" the
configuration.

Closes #1976

Change-Id: Ia1a659bd037cfd347e5bf763d7145e1523fca56a