Fix build script for receiver app.
The previous fix only fixed uncompiled mode, it turns out.
The actual root problem was a namespace collision in the build
script for the receiver app, which lead to the player namespace
being over-ridden by just debug/log.js.
This fixes the build script to just not include debug/log.js. The
receiver app wasn't using that file anyway.

Change-Id: I290a6a9051c86c816614afe3922d2c82bf8cbfa5