Fix edge case in DASH tests
During a refactor ("Replace find/get callbacks with SegmentIndex"),
this test changed from using a hard-coded segment position to looking
one up based on the live edge.  But using the exact live edge leads to
some amount of flakiness.

Back off the live edge enough to make sure you're not on the trailing
edge of the segment.

Fixes b/137286528

Change-Id: Ic4e869e8a3b1ceaf45dc14b04ea0e987bd690453